chore(rhomb): removed cap-mmc-highspeed configuration for emmc
This configuration set eMMC to support highspeed timinigs. Based on experiments
with larger size data we observed that our board eMMC is not stable with this
timings and it gives i/o error, timeout err etc. It gives inconsistent
performance. This is not reliable behaviour so we have removed this
configuration.

Signed-off-by: Android Linux Media <android.linuxmedia@gmail.com>